Monday, March 11 2024

+++ AdminLTE-3.2.0/pages/tables/delete.php
@@ -1,20 +1,23 @@
 <?php
 include 'config.php';
+if ($_SERVER["REQUEST_METHOD"] == "POST") {
 
-$id = $_GET['userid']; 
+   
+    $id = mysqli_real_escape_string($conn, $_POST['userid']);
 
+    $sql = "DELETE FROM user WHERE userid='$id'";
+    $query = mysqli_query($conn, $sql);
 
-$sql = "DELETE FROM user WHERE userid='$id'";
-$query = mysqli_query($conn, $sql);
-if ($query) {
-    echo "<script language='JavaScript'>
-           (window.alert('Data User sudah dihapus'))
-           location.href='index.php'
-           </script>";
-} else {
-    echo "<script language='JavaScript'>
-           (window.alert('Data User tidak dapat dihapus'))
-           </script>";
+    if ($query) {
+        echo "<script language='JavaScript'>
+               (window.alert('Data User sudah dihapus'))
+               location.href='index.php'
+               </script>";
+    } else {
+        echo "<script language='JavaScript'>
+               (window.alert('Data User tidak dapat dihapus'))
+               location.href='index.php'
+               </script>";
+    }
 }
-
 ?>
